+++ src/dipdup/cli.py
@@ -344,20 +344,34 @@ async def config_export(ctx: click.Context, unsafe: bool, full: bool) -> None:
 
 @config.command(name='env')
 @click.option('--output', '-o', type=str, default=None, help='Output to file instead of stdout.')
+@click.option('--unsafe', is_flag=True, help='Resolve environment variables or use default values from the config.')
+@click.option('--compose', is_flag=True, help='Output in docker-compose format.')
 @click.pass_context
 @_cli_wrapper
-async def config_env(ctx: click.Context, output: str | None) -> None:
+async def config_env(ctx: click.Context, output: str | None, unsafe: bool, compose: bool) -> None:
     """Dump environment variables used in DipDup config.
 
     If variable is not set, default value will be used.
     """
-    from dipdup.config import DipDupConfig
+    from dipdup.yaml import DipDupYAMLConfig
 
-    config = DipDupConfig.load(
+    config, environment = DipDupYAMLConfig.load(
         paths=ctx.obj.config._paths,
-        environment=True,
+        environment=unsafe,
     )
-    content = '\n'.join(f'{k}={v}' for k, v in sorted(config._environment.items()))
+    if compose:
+        content = '\nservices:\n  dipdup:\n    environment:\n'
+        _tab = ' ' * 6
+        for k, v in sorted(environment.items()):
+            line = f'{_tab}- {k}=' + '${' + k
+            if v:
+                line += ':-' + v + '}'
+            else:
+                line += '}'
+
+            content += line + '\n'
+    else:
+        content = '\n'.join(f'{k}={v}' for k, v in sorted(environment.items()))
     if output:
         Path(output).write_text(content)
     else: