+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog], and this project adheres to [Semantic Versioning].
 
+## [Unreleased]
+
+### Added
+
+- cli: Added `--unsafe` and `--compose` flags to `config env` command.
+
 ## [7.0.0] - 2023-09-25
 
 ### Fixed